Add a keys operation to keyvalue.widl
Redis supports a `keys` method that returns a list of keys that specify a given query.  For instance, `connection.keys("*")` will return all keys in the store.

It's possible we need to paginate this method in a similar fashion to the `Range` operation.  I'm happy to update the PR accordingly.  My current use case is small so retrieving all keys in the store is safe.